+++ cmd/plumb/main_test.go
@@ -1,4 +1,3 @@
-// cmd/plumb/main_test.go
 package main
 
 import (